## Motivation

Add support for [NCCL symmetric memory](https://docs.nvidia.com/deeplearning/nccl/user-guide/docs/usage/bufferreg.html#window-registration).
This PR only enables TP allreduce. Follow up will enable other collectives.

NCCL symmetric memory uses a separate memory pool to allocate tensors that will be used for communications, therefore it can lead to more memory fragmentation. This new feature is off by default and can be turned on with `--enable-symm-mem` server argument.

## FP8
Baseline:
```
SGLANG_CUTLASS_MOE=1 SGLANG_ENABLE_FLASHINFER_GEMM=1 python3 -m sglang.launch_server --tp=8 --trust-remote-code --disable-radix-cache --model-path=deepseek-ai/DeepSeek-R1-0528 --attention-backend=cutlass_mla
python3 -m sglang.bench_serving --backend sglang --dataset-name random --num-prompt 512 --random-input 1024 --random-output 1024 --random-range-ratio 1 --max-concurrency 512 --warmup-requests 512
```
`6671.94 tok/s`

with this PR:
```
SGLANG_CUTLASS_MOE=1 SGLANG_ENABLE_FLASHINFER_GEMM=1 python3 -m sglang.launch_server --tp=8 --trust-remote-code --disable-radix-cache --model-path=deepseek-ai/DeepSeek-R1-0528 --attention-backend=cutlass_mla --enable-symm-mem
python3 -m sglang.bench_serving --backend sglang --dataset-name random --num-prompt 512 --random-input 1024 --random-output 1024 --random-range-ratio 1 --max-concurrency 512 --warmup-requests 512
```
`7129.24 tok/s`

## FP4

### Performance with concurrency of `256`:
Baseline:
```
python3 -m sglang.launch_server --model-path nvidia/DeepSeek-R1-FP4 --trust-remote-code --quantization modelopt_fp4 --tp 8 --enable-flashinfer-cutlass-moe  --enable-ep-moe --ep-size 8 --max-running-requests 256 --disable-radix-cache --mem-fraction-static 0.9 --stream-output --cuda-graph-bs 256  --cuda-graph-max-bs 256
python3 -m sglang.bench_serving --backend sglang --dataset-name random --num-prompt 256 --random-input 1024 --random-output 1024 --random-range-ratio 1 --max-concurrency 256
```
`5286.31 tok/s`

with this PR:
```
python3 -m sglang.launch_server --model-path nvidia/DeepSeek-R1-FP4 --trust-remote-code --quantization modelopt_fp4 --tp 8 --enable-flashinfer-cutlass-moe  --enable-ep-moe --ep-size 8 --max-running-requests 256 --disable-radix-cache --mem-fraction-static 0.9 --stream-output --cuda-graph-bs 256  --cuda-graph-max-bs 256 --enable-symm-mem
python3 -m sglang.bench_serving --backend sglang --dataset-name random --num-prompt 256 --random-input 1024 --random-output 1024 --random-range-ratio 1 --max-concurrency 256
```
`5518.76 tok/s`
### Performance with concurrency of `128`:
Baseline:
```
python3 -m sglang.launch_server --model-path nvidia/DeepSeek-R1-FP4 --trust-remote-code --quantization modelopt_fp4 --tp 8 --enable-flashinfer-cutlass-moe  --enable-ep-moe --ep-size 8 --max-running-requests 128 --disable-radix-cache --mem-fraction-static 0.9 --stream-output --cuda-graph-bs 128  --cuda-graph-max-bs 128
python3 -m sglang.bench_serving --backend sglang --dataset-name random --num-prompt 128 --random-input 1024 --random-output 1024 --random-range-ratio 1 --max-concurrency 128
```
`3632.57 tok/s`

with this PR:
```
python3 -m sglang.launch_server --model-path nvidia/DeepSeek-R1-FP4 --trust-remote-code --quantization modelopt_fp4 --tp 8 --enable-flashinfer-cutlass-moe  --enable-ep-moe --ep-size 8 --max-running-requests 128 --disable-radix-cache --mem-fraction-static 0.9 --stream-output --cuda-graph-bs 128  --cuda-graph-max-bs 128 --enable-symm-mem
python3 -m sglang.bench_serving --backend sglang --dataset-name random --num-prompt 128 --random-input 1024 --random-output 1024 --random-range-ratio 1 --max-concurrency 128
```
`3713.95 tok/s`
## Modifications

<!-- Describe the changes made in this PR. -->

## Checklist

- [x] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [x] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [x] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.